requirements.txt
# Core numeric
numpy
scipy
pytorch-msssim

# Visualization
matplotlib
plotly
open3d
Pillow
plyfile
OpenEXR

# Utilities
einops
tqdm
pyyaml
packaging
pytz

# Config / infra
hydra-core
omegaconf
iopath
timm
decord
av
pycocotools

# Hugging Face
huggingface_hub>=0.20.0

gsplat
